demos/project-stories/requirements.txt
django>=3.2,<4
django-avatar>=5.0,<5.1
django-contrib-comments>=2.1,<2.2
django-rosetta>=0.9,<0.10
djangorestframework>=3.12,<3.13
six
